// client/types/reply.ts
export type ReplyState = {
    author: {
        displayName: string;
        email: string;
        photoURL: string;
        uid: string;
        userNum: number;
        __v: number;
        _id: string;
    };
    createdAt: string;
    postId: string;
    reply: string;
    updatedAt: string;
    __v: number;
    _id: string;
};